requirements.txt
Babel==0.9.6
BeautifulSoup==3.2.1
Mako==0.8.1
Twisted==13.2.0
https://ipaddr-py.googlecode.com/files/ipaddr-2.1.10.tar.gz#sha1=c608450b077b19773d4f1b5f1ef88b26f6650ce0#egg=ipaddr-2.1.10-py2.7
pycrypto==2.6.1
pyOpenSSL==0.14
pygeoip==0.2.7
gnupg==2.0.1
pillow>=2.6.1
qrcode==5.0.1
service_identity==14.0.0
stem==1.3.0
zope.interface==3.6.1
